move AuthTokenManager to rundeck-app-util library, remove rundeck-data-spi
change package name
update build config
