+++ rundeck-app-util/src/main/java/org/rundeck/app/util/spi/AuthTokenManager.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.dtolabs.rundeck.core.authentication.tokens;
+package org.rundeck.app.util.spi;
 
 import com.dtolabs.rundeck.core.authorization.UserAndRolesAuthContext;
 import org.rundeck.app.data.model.v1.AuthenticationToken;
