+++ rundeckapp/build.gradle
@@ -76,7 +76,7 @@ dependencies {
     implementation project(':core')
     implementation project(':rundeck-storage:rundeck-storage-filesys')
     implementation project(':rundeck-data-model')
-    implementation project(':rundeck-data-spi')
+    implementation project(':rundeck-app-util')
 
 
     /** It's important that the configuration addition follow the form 'compile project("")'
