+++ settings.gradle
@@ -27,8 +27,8 @@ include	'rundeck-storage',
     'rundeck-authz:rundeck-authz-api',
     'rundeck-authz:rundeck-authz-core',
     'rundeck-authz:rundeck-authz-yaml',
-    'rundeck-data-spi',
     'rundeck-data-model',
+    'rundeck-app-util',
     'core',
     'plugins:localexec-plugin',
     'plugins:script-plugin',
